# artifacts/api-server/src/routes/properties.ts
import { Router, type IRouter } from "express";
import { eq } from "drizzle-orm";
import { db, propertiesTable } from "@workspace/db";
import { ListPropertiesResponse, GetPropertyParams, GetPropertyResponse } from "@workspace/api-zod";

const router: IRouter = Router();

router.get("/properties", async (_req, res): Promise<void> => {
  const rows = await db.select().from(propertiesTable).where(eq(propertiesTable.isActive, true));
  res.json(ListPropertiesResponse.parse(rows));
});

router.get("/properties/:id", async (req, res): Promise<void> => {
  const params = GetPropertyParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const [row] = await db
    .select()
    .from(propertiesTable)
    .where(eq(propertiesTable.id, params.data.id))
    .limit(1);
  if (!row) {
    res.status(404).json({ error: "Property not found" });
    return;
  }
  res.json(GetPropertyResponse.parse(row));
});

export default router;