# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChaletDeLaPointe reservation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type SeasonName = (typeof SeasonName)[keyof typeof SeasonName];

export const SeasonName = {
  haute: "haute",
  mi: "mi",
  basse: "basse",
} as const;

export type UserGroup = (typeof UserGroup)[keyof typeof UserGroup];

export const UserGroup = {
  proprietaire: "proprietaire",
  ancien_proprietaire: "ancien_proprietaire",
  famille: "famille",
  ami: "ami",
  public: "public",
} as const;

export type ReservationStatus =
  (typeof ReservationStatus)[keyof typeof ReservationStatus];

export const ReservationStatus = {
  draft: "draft",
  pending_payment: "pending_payment",
  confirmed: "confirmed",
  cancelled: "cancelled",
  expired: "expired",
} as const;

export interface Property {
  id: string;
  slug: string;
  name: string;
  description: string;
  city: string;
  province: string;
  country: string;
  bedrooms: number;
  bathrooms: number;
  maxGuests: number;
  isActive: boolean;
}

export interface QuoteRequest {
  propertyId: string;
  startDate: string;
  endDate: string;
  /** @minimum 1 */
  guests: number;
}

export interface SeasonBreakdownEntry {
  seasonName: SeasonName;
  nights: number;
  basePricePerNight: number;
  effectivePricePerNight: number;
  subtotal: number;
}

export interface Quote {
  propertyId: string;
  startDate: string;
  endDate: string;
  nights: number;
  subtotal: number;
  discountAmount: number;
  total: number;
  appliedGroup: UserGroup;
  seasonBreakdown: SeasonBreakdownEntry[];
}

export interface CreateReservationInput {
  propertyId: string;
  startDate: string;
  endDate: string;
  /** @minimum 1 */
  guests: number;
  /** @nullable */
  notes?: string | null;
}

export interface Reservation {
  id: string;
  propertyId: string;
  userId: string;
  startDate: string;
  endDate: string;
  guests: number;
  nights: number;
  subtotal: number;
  discountAmount: number;
  total: number;
  status: ReservationStatus;
  /** @nullable */
  notes?: string | null;
  seasonBreakdown: SeasonBreakdownEntry[];
  createdAt: string;
  updatedAt: string;
}

export interface CalendarEntry {
  propertyId: string;
  startDate: string;
  endDate: string;
}

export interface AdminActionBody {
  /** @nullable */
  reason?: string | null;
}

export interface AdminModifyBody {
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  /**
   * @minimum 1
   * @nullable
   */
  guests?: number | null;
  status?: ReservationStatus | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  reason?: string | null;
}