# lib/api-zod/src/generated/types/calendarEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChaletDeLaPointe reservation API
 * OpenAPI spec version: 0.1.0
 */

export interface CalendarEntry {
  propertyId: string;
  startDate: Date;
  endDate: Date;
}